requirements.txt
ansible-core==2.17.0
yamllint==1.35.1
ansible-lint==24.2.3

